move algorithm macros into separate module
